requirements.txt
biopython==1.79
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
dgl==0.6.1
edlib==1.3.8.post2
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
mappy==2.20
matplotlib==3.4.2
networkx==2.5.1
numpy==1.20.3
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
threadpoolctl==2.1.0
torch==1.8.1+cpu
torchaudio==0.8.1
torchvision==0.9.1+cpu
typing-extensions==3.10.0.0
urllib3==1.26.5
